Build skeleton for custom notifications
Custom notifications find the largest "cell" of 8 surrounding the
currently focused window. The notification is placed in the center.

Still need to make text and display/deletion depend upon modal entry
and exit. Implement this in the sequentialKeys spoon, likely in the
overloaded entry and exit functions. Also need to set up a window
watcher that changes the cell in which the notification is displayed
if the window is resized, since some keybindings to hyper mode work
with a semi-persistent mode (e.g. windows switching doesn't auto-exit
from hyper mode).