Replace CMD-Z with Key Event Destruction
Using sequential modes, it's common to press a key by accident. In
an editor, this often types an unwanted character. When debugging
Hammerspoon, this is the most frequent source of bugs. To avoid this,
I used hs.eventtap to watch for key releases, then key stroke CMD-Z if
the key was a printable character. This new method is safer and
cleaner.