De-hardcode shift margins
Shift margins were hardcoded into the windows manipulation spoon. Now
they are only set to a non-(0,0) value outside of the spoon. If the
focused screen isn't a laptop screen, any values are temporarily reset
to (0,0). Note that this only works for one screen at a time. To
use layouts, you must set the value for one screen, build the layout,
set the value for the second screen, build the second layout, then
concatenate the two layouts into a single layout to be applied.